# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/audit-logs/entity/{entityType}/{entityId}": {
      "get": {
        "description": "Retourne tous les changements enregistrés pour une entité (SalesOrder, Invoice, ProductionOrder, etc.)",
        "operationId": "AuditController_getEntityLogs",
        "parameters": [
          {
            "name": "entityType",
            "required": true,
            "in": "path",
            "description": "Type d’entité (ex. SalesOrder, Invoice)",
            "schema": {
              "example": "SalesOrder",
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": true,
            "in": "path",
            "description": "Identifiant de l’entité",
            "schema": {
              "example": "so123",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre max de logs retournés",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 50,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit logs de l'entité",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Historique d'audit d'une entité",
        "tags": [
          "Audit"
        ]
      }
    },
    "/audit-logs/user/{userId}": {
      "get": {
        "description": "Retourne toutes les actions qu'a effectuées un utilisateur",
        "operationId": "AuditController_getUserLogs",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Identifiant utilisateur",
            "schema": {
              "example": "usr123",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre max de logs retournés",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 50,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit logs de l'utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actions d'un utilisateur",
        "tags": [
          "Audit"
        ]
      }
    },
    "/audit-logs": {
      "get": {
        "description": "Retourne tous les audit logs avec filtres optionnels. Requiert le rôle ADMIN.",
        "operationId": "AuditController_getAllLogs",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre max de logs retournés",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 50,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "description": "Filtrer par type d’entité",
            "schema": {
              "example": "SalesOrder",
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "description": "Filtrer par action (enum AuditAction)",
            "schema": {
              "example": "SALES_ORDER_CREATED",
              "type": "string",
              "enum": [
                "SALES_ORDER_CREATED",
                "SALES_ORDER_STATUS_CHANGED",
                "SALES_ORDER_BAT_SENT",
                "SALES_ORDER_BAT_APPROVED",
                "INVOICE_CREATED",
                "INVOICE_PAYMENT_RECORDED",
                "INVOICE_MARKED_PAID",
                "PRODUCTION_ORDER_CREATED",
                "PRODUCTION_ORDER_PROGRESS_UPDATED",
                "PRODUCTION_ORDER_QUALITY_APPROVED",
                "PURCHASE_ORDER_CREATED",
                "PURCHASE_ORDER_STATUS_CHANGED",
                "DELIVERY_CREATED",
                "DELIVERY_MARKED_DELIVERED"
              ]
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filtrer par utilisateur",
            "schema": {
              "example": "usr123",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Début de période (ISO)",
            "schema": {
              "example": "2026-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Fin de période (ISO)",
            "schema": {
              "example": "2026-12-31T23:59:59.999Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tous les audit logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Journal d'audit complet (admin)",
        "tags": [
          "Audit"
        ]
      }
    },
...
  },
  "info": {
    "title": "Raffia Weave Flow API",
    "description": "Fragment OpenAPI des modules contractualisés (search/reference-lookup/audit/templates/images). Document complet: GET /{SWAGGER_PATH}-json sur l’API vivante.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AuditLogUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "usr123"
          },
          "email": {
            "type": "string",
            "example": "admin@raffia.test"
          },
          "name": {
            "type": "object",
            "example": "Admin Raffia",
            "nullable": true
          },
          "role": {
            "type": "string",
            "description": "Présent sur les listes entity / admin, absent sur user/:userId",
            "enum": [
              "GERANT",
              "RESPONSABLE_GENERAL",
              "RESPONSABLE_PRODUCTION",
              "RESPONSABLE_LIVRAISON",
              "RESPONSABLE_FINANCIER_STOCKS"
            ],
            "example": "GERANT"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "AuditLogResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "aud123"
          },
          "entityType": {
            "type": "string",
            "description": "Type d’entité auditée",
            "example": "SalesOrder"
          },
          "entityId": {
            "type": "string",
            "example": "so123"
          },
          "action": {
            "type": "string",
            "description": "Action auditée (enum AuditAction Prisma)",
            "example": "SALES_ORDER_STATUS_CHANGED"
          },
          "userId": {
            "type": "string",
            "example": "usr123"
          },
          "user": {
            "description": "Utilisateur à l’origine de l’action",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuditLogUserResponseDto"
              }
            ]
          },
          "changes": {
            "type": "object",
            "description": "Diff JSON des champs modifiés (before/after)",
            "nullable": true,
            "additionalProperties": true,
            "example": {
              "before": {
                "status": "TO_PROCESS"
              },
              "after": {
                "status": "IN_PRODUCTION"
              }
            }
          },
          "details": {
            "type": "object",
            "description": "Détail libre (raison, référence…)",
            "nullable": true,
            "example": "Passage en production"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "entityType",
          "entityId",
          "action",
          "userId",
          "user",
          "createdAt"
        ]
      },
...
    }
  }
}